requirements.txt
opencv-python
numpy
dlib
face-utils
imutils
pushbullet.py
scipy